[INFRA] Chop chopper into libraries
