+++ include/chopper/detail_starts_with.hpp
@@ -3,7 +3,7 @@
 #include <string>
 #include <string_view>
 
-bool starts_with(std::string const & target, std::string_view const & query)
+inline bool starts_with(std::string const & target, std::string_view const & query)
 {
     size_t index{};
     while (index < target.size() && index < query.size() && target[index] == query[index])
