+++ src/chopper_search.cpp
@@ -1,0 +1,128 @@
+#include <fstream>
+#include <seqan3/std/ranges>
+
+#include <cereal/archives/binary.hpp>
+
+#include <seqan3/argument_parser/all.hpp>
+#include <seqan3/io/sequence_file/input.hpp>
+#include <seqan3/core/debug_stream.hpp>
+#include <seqan3/range/views/interleave.hpp>
+#include <seqan3/range/views/kmer_hash.hpp>
+
+#include <chopper/search/search.hpp>
+#include <chopper/search/search_config.hpp>
+#include <chopper/search/search_data.hpp>
+
+void initialize_argument_parser(seqan3::argument_parser & parser, search_config & config)
+{
+    parser.info.author = "Avenja";
+    parser.info.short_description = "Read an HIBF on results from chopper-build and search queries in it.";
+    parser.info.version = "1.0.0";
+
+    // todo: k-mer size should be serialized with building the index to avoid inconsistencies
+    parser.add_option(config.chopper_index_filename, 'i', "index", "Provide the HIBF index file produced by chopper build.");
+    parser.add_option(config.k, 'k', "kmer-size", "The kmer size to build kmers.");
+    parser.add_option(config.errors, 'e', "errors", "The errors to allow in the search.");
+    parser.add_option(config.query_filename, 'q', "queries", "The query sequences to seach for in the index.");
+    parser.add_flag(config.verbose, 'v', "verbose", "Output logging/progress information.");
+}
+
+struct search_file_traits : public seqan3::sequence_file_input_default_traits_dna
+{
+    using sequence_alphabet = seqan3::dna4;
+};
+
+using search_sequence_file_t = seqan3::sequence_file_input<search_file_traits,
+                                                    seqan3::fields<seqan3::field::id, seqan3::field::seq>,
+                                                    seqan3::type_list<seqan3::format_fasta, seqan3::format_fastq>>;
+
+std::vector<size_t> compute_kmers(seqan3::dna4_vector const & query, search_config const & config)
+{
+    std::vector<size_t> kmers{};
+
+    for (auto hash : query | seqan3::views::kmer_hash(seqan3::ungapped{config.k}))
+        kmers.push_back(hash);
+
+    return kmers;
+}
+
+void write_header(search_data const & data, std::ostream & out_stream)
+{
+    data.user_bins.write_filenames(out_stream);
+    out_stream << "#QUERY_NAME\tUSER_BINS\n";
+}
+
+void write_result(std::unordered_set<std::pair<int32_t, uint32_t>, pair_hash> const & membership_result,
+                  std::string const & id,
+                  search_data const & data,
+                  std::ostream & out_stream)
+{
+    if (membership_result.empty())
+    {
+        out_stream << id << '\t' << std::endl;
+        return;
+    }
+
+    // storing and sorting this is only done for testing purposes.
+    // If this turns out to have a significant runtime penalty, it should be removed.
+    std::vector<int64_t> result_positions; // TODO allocate this outside of this function
+    for (auto const & [ibf_idx, bin_idx] : membership_result)
+    {
+        assert(data.user_bins.get_position(ibf_idx, bin_idx) > -1);
+        result_positions.push_back(data.user_bins.get_position(ibf_idx, bin_idx));
+    }
+    std::sort(result_positions.begin(), result_positions.end()); // otherwise the result output is not testable
+
+    out_stream << id << '\t';
+    for (size_t i = 0; i < result_positions.size() - 1; ++i)
+        out_stream << result_positions[i] << ',';
+    out_stream << result_positions.back() << std::endl;
+}
+
+int chopper_search(seqan3::argument_parser & parser)
+{
+    search_config config{};
+    initialize_argument_parser(parser, config);
+
+    try
+    {
+        parser.parse();
+    }
+    catch (seqan3::argument_parser_error const & ext)
+    {
+        seqan3::debug_stream << "[CHOPPER SEARCH ERROR] " << ext.what() << "\n";
+        return -1;
+    }
+
+    search_data data;
+
+    { // read ibf vector
+        std::ifstream is{config.chopper_index_filename, std::ios::binary};
+
+        if (!is.good() && !is.is_open())
+            throw std::runtime_error{"File " + config.chopper_index_filename + " could not be opened"};
+
+        cereal::BinaryInputArchive iarchive{is};
+        iarchive(data.hibf);
+        iarchive(data.hibf_bin_levels);
+        iarchive(data.user_bins);
+    }
+
+    write_header(data, std::cout);
+
+    std::vector<size_t> read_kmers; // allocate space once
+    std::unordered_set<std::pair<int32_t, uint32_t>, pair_hash> result{};
+
+    for (auto && [id, seq] : search_sequence_file_t{config.query_filename})
+    {
+        read_kmers = compute_kmers(seq, config);
+        result.clear();
+
+        search(result, read_kmers, data, config, 0); // start at top level ibf
+
+        write_result(result, id, data, std::cout);
+    }
+
+    return 0;
+}
+
